Handing some of the problems exposed by Coverity
Mostly adding strncpy termination to some components (coretemp, lmsensors, micpower).
Removed some unused component writing functions (lustre, mx).
Fixed CUDA component configure.in to get the correct version of nvcc.
Fixed division so it works in double precision rather than integer in the rapl component.
Fixed a minor complaint about a stack counter variable in papi_preset.

Thanks to William Cohen for sending the Coverity results report.

+++ src/components/rapl/tests/rapl_basic.c
@@ -291,7 +291,7 @@ int main (int argc, char **argv)
 		  	}
 		}
 	}
-	max_time = elapsed_time * (0xffffffff / max_value);
+	max_time = elapsed_time * ( (double)0xffffffff / (double)max_value );
 	printf("\n");
 	printf ("Approximate time to energy measurement wraparound: %.3f sec or %.3f min.\n", 
 		max_time, max_time/60);
